chore(RockCrafter): introduce abstract base class
